Fix StyleCop warnings in DataDrivenTest\Number -TestDecimalAndThousandsSeparators -LongFormTestConfiguration -TestNumberRecognizerCache -TestNumberRecognizerInitialization -TestParserFactory -TestNumber_Portuguese -TestNumber_Spanish -TestNumber_Korean -TestNumber_Japanese -TestNumber_Italian -TestNumber_German -TestNumber_French -TestNumber_English -TestNumber_Dutch -TestNumber_Chinese
