+++ .NET/Microsoft.Recognizers.Text.DataDrivenTests/Number/TestDecimalAndThousandsSeparators.cs
@@ -8,70 +8,6 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Recognizers.Text.Number.Tests
 {
-    public class LongFormTestConfiguration : INumberParserConfiguration
-    {
-        public ImmutableDictionary<string, long> CardinalNumberMap { get; }
-
-        public ImmutableDictionary<string, long> OrdinalNumberMap { get; }
-
-        public ImmutableDictionary<string, long> RoundNumberMap { get; }
-
-        public NumberOptions Options { get; }
-
-        public CultureInfo CultureInfo { get; }
-
-        public Regex DigitalNumberRegex { get; }
-
-        public Regex FractionPrepositionRegex { get; }
-
-        public string FractionMarkerToken { get; }
-
-        public Regex HalfADozenRegex { get; }
-
-        public string HalfADozenText { get; }
-
-        public string LangMarker { get; } = "SelfDefined";
-
-        public char NonDecimalSeparatorChar { get; }
-
-        public char DecimalSeparatorChar { get; }
-
-        public string WordSeparatorToken { get; }
-
-        public IEnumerable<string> WrittenDecimalSeparatorTexts { get; }
-
-        public IEnumerable<string> WrittenGroupSeparatorTexts { get; }
-
-        public IEnumerable<string> WrittenIntegerSeparatorTexts { get; }
-
-        public IEnumerable<string> WrittenFractionSeparatorTexts { get; }
-
-        // Test-specific initialization: the Regex matches nothing.
-        public Regex NegativeNumberSignRegex { get; } = new Regex(@"[^\s\S]"); 
-
-        public LongFormTestConfiguration(char decimalSep, char nonDecimalSep)
-        {
-            this.DecimalSeparatorChar = decimalSep;
-            this.NonDecimalSeparatorChar = nonDecimalSep;
-            this.CultureInfo = new CultureInfo(Culture.English);
-            this.CardinalNumberMap = ImmutableDictionary<string, long>.Empty;
-            this.OrdinalNumberMap = ImmutableDictionary<string, long>.Empty;
-            this.RoundNumberMap = ImmutableDictionary<string, long>.Empty;
-            this.DigitalNumberRegex = new Regex(@"((?<=\b)(hundred|thousand|million|billion|trillion|dozen(s)?)(?=\b))|((?<=(\d|\b))(k|t|m|g|b)(?=\b))",
-                                                RegexOptions.Singleline);
-        }
-
-        public IEnumerable<string> NormalizeTokenSet(IEnumerable<string> tokens, ParseResult context)
-        {
-            throw new NotImplementedException();
-        }
-
-        public long ResolveCompositeNumber(string numberStr)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
     [TestClass]
     public class TestDecimalAndThousandsSeparators
     {
@@ -79,11 +15,12 @@ namespace Microsoft.Recognizers.Text.Number.Tests
         {
             char decimalSep = type.DecimalsMark, nonDecimalSep = type.ThousandsMark;
 
-            var parser = AgnosticNumberParserFactory.GetParser(AgnosticNumberParserType.Double,
-                                                               new LongFormTestConfiguration(decimalSep, nonDecimalSep));
+            var parser = AgnosticNumberParserFactory.GetParser(
+                AgnosticNumberParserType.Double, new LongFormTestConfiguration(decimalSep, nonDecimalSep));
 
-            var resultJson = parser.Parse(new ExtractResult() {
-                Text = query, Start = 0, Length = query.Length, Type = "builtin.num.double", Data = "Num"
+            var resultJson = parser.Parse(new ExtractResult()
+            {
+                Text = query, Start = 0, Length = query.Length, Type = "builtin.num.double", Data = "Num",
             });
 
             Assert.AreEqual(value, resultJson.ResolutionStr);
